Extracting autocomplete to own file and adding dev portal support
